% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasearch.R
\name{datasearch}
\alias{datasearch}
\title{Summarize Available Datasets}
\usage{
datasearch(pkgs = NULL, filter = NULL)
}
\arguments{
\item{pkgs}{a vector of package names, if \code{NULL} all installed packages
are considered.}

\item{filter}{a predicate function, or formula, to be applied on each dataset
to restrict the output}
}
\description{
Summarize Available Datasets
}
